- refactor three-sum algorithm to be in a single file - add unit tests for three-sum algorithm - refactor three-sum and stop-watch clients
