+++ src/algorithms/index.js
@@ -2,5 +2,6 @@ module.exports = {
   BinarySearch: require('./binary-search/binary-search'),
   DijkstraEvaluate: require('./dijkstra-two-stack/dijkstra-two-stack'),
   Euclidean: require('./euclidean/euclidean'),
-  LinearSearch: require('./linear-search/linear-search')
+  LinearSearch: require('./linear-search/linear-search'),
+  ThreeSum: require('./three-sum/three-sum')
 }
